rename business logic to utils
