+++ src/components/FuelSavingsResults.js
@@ -1,5 +1,5 @@
 import React, {PropTypes} from 'react';
-import NumberFormatter from '../businessLogic/numberFormatter';
+import NumberFormatter from '../utils/numberFormatter';
 
 //This is a stateless functional component. (Also known as pure or dumb component)
 //More info: https://facebook.github.io/react/blog/2015/10/07/react-v0.14.html#stateless-functional-components
