+++ src/store/store.spec.js
@@ -2,8 +2,8 @@ import * as ActionTypes from '../constants/actionTypes';
 import { createStore } from 'redux';
 import { expect } from 'chai';
 import rootReducer from '../reducers';
-import calculator from '../businessLogic/fuelSavingsCalculator';
-import dateHelper from '../businessLogic/dateHelper';
+import calculator from '../utils/fuelSavingsCalculator';
+import dateHelper from '../utils/dateHelper';
 
 const initialState = {
   fuelSavingsAppState: {
