+++ src/utils/fuelSavingsCalculator.js
