+++ src/utils/numberFormatter.spec.js
